+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -14,40 +14,53 @@ import roomescape.reservation.dto.response.AdminReservationPageResponse.AdminRes
 import roomescape.reservation.dto.response.CreateReservationResponse;
 import roomescape.reservation.repository.ReservationDao;
 import roomescape.reservationtime.domain.ReservationTime;
-import roomescape.reservationtime.repository.ReservationTimeDao;
+import roomescape.reservationtime.service.ReservationTimeService;
 import roomescape.theme.domain.Theme;
-import roomescape.theme.repository.ThemeDao;
+import roomescape.theme.service.ThemeService;
 import roomescape.user.domain.User;
-import roomescape.user.repository.UserDao;
+import roomescape.user.service.UserService;
 
 @Service
 public class ReservationService {
 
     private final ReservationDao reservationDao;
-    private final ReservationTimeDao reservationTimeDao;
-    private final ThemeDao themeDao;
-    private final UserDao userDao;
+    private final UserService userService;
+    private final ReservationTimeService reservationTimeService;
+    private final ThemeService themeService;
 
     public ReservationService(ReservationDao reservationDao,
-                              ReservationTimeDao reservationTimeDao,
-                              ThemeDao themeDao, UserDao userDao) {
+                              UserService userService,
+                              ReservationTimeService reservationTimeService,
+                              ThemeService themeService) {
         this.reservationDao = reservationDao;
-        this.reservationTimeDao = reservationTimeDao;
-        this.themeDao = themeDao;
-        this.userDao = userDao;
+        this.userService = userService;
+        this.reservationTimeService = reservationTimeService;
+        this.themeService = themeService;
     }
 
     public CreateReservationResponse addReservation(CreateReservationRequest createReservationRequest) {
-        ReservationTime reservationTime = reservationTimeDao.findById(createReservationRequest.timeId())
-                .orElseThrow(() -> new ResourceNotFoundException("존재하지 않는 시간입니다."));
-        Theme theme = themeDao.findById(createReservationRequest.themeId())
-                .orElseThrow(() -> new ResourceNotFoundException("존재하지 않는 테마입니다."));
-        User user = userDao.findById(createReservationRequest.memberId())
-                .orElseThrow(() -> new ResourceNotFoundException("존재하지 않는 사용자입니다."));
-        LocalDate reservationDate = createReservationRequest.date();
-        if (!reservationDate.isAfter(LocalDate.now())) {
+        ReservationTime reservationTime = reservationTimeService
+                .getReservationTimeById(createReservationRequest.timeId());
+        Theme theme = themeService.getThemeById(createReservationRequest.themeId());
+        User user = userService.getUserById(createReservationRequest.memberId());
+
+        validateNewReservation(createReservationRequest);
+
+        Reservation reservation = new Reservation(
+                user,
+                createReservationRequest.date(),
+                reservationTime,
+                theme
+        );
+        Reservation saved = reservationDao.save(reservation);
+        return CreateReservationResponse.from(saved);
+    }
+
+    private void validateNewReservation(CreateReservationRequest createReservationRequest) {
+        if (!createReservationRequest.date().isAfter(LocalDate.now())) {
             throw new BadRequestException("하루 전 까지 예약 가능합니다.");
         }
+
         if (reservationDao.isExistByThemeIdAndTimeIdAndDate(
                 createReservationRequest.themeId(),
                 createReservationRequest.timeId(),
@@ -55,15 +68,6 @@ public class ReservationService {
         ) {
             throw new BadRequestException("이미 해당 시간에 예약이 존재합니다.");
         }
-        Reservation reservation = new Reservation(
-                null,
-                user,
-                createReservationRequest.date(),
-                reservationTime,
-                theme
-        );
-        Reservation savedReservation = reservationDao.save(reservation);
-        return CreateReservationResponse.from(savedReservation);
     }
 
     public void deleteReservation(Long id) {
